% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel_classification.R
\name{add_bagel_classifications}
\alias{add_bagel_classifications}
\title{Add BAGEL classifications}
\usage{
add_bagel_classifications(
  data = NULL,
  gene_column = 1,
  ess = NULL,
  noness = NULL
)
}
\arguments{
\item{data}{data frame.}

\item{gene_column}{the index of column containing gene symbols.}

\item{ess}{vector of essential gene symbols.}

\item{noness}{vector of non-essential gene symbols.}
}
\value{
logical.
}
\description{
Add BAGEL classifications to data frame.
}
