% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{add_pseudocount}
\alias{add_pseudocount}
\title{Add pseudocount}
\usage{
add_pseudocount(data = NULL, pseudocount = 5, indices = NULL, ...)
}
\arguments{
\item{data}{a data frame.}

\item{pseudocount}{a pseudocount (Default: 5).}

\item{indices}{column indices.}

\item{...}{parameters for \code{check_dataframe}.}
}
\value{
a data frame.
}
\description{
Add pseudocount to selected dataframe columns
}
