% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel_classification.R
\name{average_guide_bfs}
\alias{average_guide_bfs}
\title{Get average sgRNA BF per gene}
\usage{
average_guide_bfs(data = NULL, gene_column = NULL, data_columns = NULL)
}
\arguments{
\item{data}{data frame}

\item{gene_column}{index of column for genes}

\item{data_columns}{index of columns to average}
}
\description{
Calculate average of guide Bayes factors per gene
}
