% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{average_replicates}
\alias{average_replicates}
\title{Average replicates}
\usage{
average_replicates(data = NULL, gene_column = NULL, data_columns = NULL)
}
\arguments{
\item{data}{data frame}

\item{gene_column}{index of column for rownames}

\item{data_columns}{index of columns to average}
}
\description{
Calculate average of replicates
}
