% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{bagel_gene_infile_options}
\alias{bagel_gene_infile_options}
\title{Optparse options for BAGEL}
\usage{
bagel_gene_infile_options()
}
\value{
list of optparse options
}
\description{
Optparse options for BAGEL
}
