% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{bagel_normalisation_options}
\alias{bagel_normalisation_options}
\title{Optparse options for output paths}
\usage{
bagel_normalisation_options()
}
\value{
list of optparse options
}
\description{
Optparse options for output paths
}
