% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_counts.R
\name{bagel_normalise_counts}
\alias{bagel_normalise_counts}
\title{Normalise counts using BAGEL method.}
\usage{
bagel_normalise_counts(
  sample_counts = NULL,
  scaling_factor = 1e+07,
  pseudocount = 5
)
}
\arguments{
\item{sample_counts}{a sample count matrix.}

\item{scaling_factor}{a scaling factor (Default: 10000000).}

\item{pseudocount}{a pseudocount (Default: 5).}
}
\value{
dataframe
}
\description{
Normalise count matrix using BAGEL method.
}
\details{
Assumes that sgRNA identifiers and gene names are in the first two columns.
First a defined \code{pseudocount} gets added to all guide counts. Then sample counts
are total normalised and multiplied by a scaling factor.
}
