% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{basic_infile_options}
\alias{basic_infile_options}
\title{Optparse options for basic input}
\usage{
basic_infile_options()
}
\value{
list of optparse options
}
\description{
Optparse options for basic input
}
