% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{basic_outfile_options}
\alias{basic_outfile_options}
\title{Optparse options for basic output}
\usage{
basic_outfile_options()
}
\value{
list of optparse options
}
\description{
Optparse options for basic output
}
