% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_changes.R
\name{calculate_gene_lfc}
\alias{calculate_gene_lfc}
\title{Calculate gene log fold changes.}
\usage{
calculate_gene_lfc(
  data = NULL,
  id_column = 1,
  gene_column = 2,
  sample_columns = NULL
)
}
\arguments{
\item{data}{sample sgRNA fold change matrix.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{sample_columns}{the index of columns containing sample data to process.}
}
\value{
a data frame containing sample gene-level log fold changes.
}
\description{
Calculate gene log fold changes.
}
\details{
Takes an sgRNA fold change matrix and averages by gene.

Requires column indices be defined:
\itemize{
\item \code{id_column} - column containing guide (sgRNA) identifiers (Default = 1).
\item \code{gene_column} - column containing gene symbols/identifiers (Default = 2).
}
}
