% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_count_statistics.R
\name{calculate_gini_index}
\alias{calculate_gini_index}
\title{Calculate gini index}
\usage{
calculate_gini_index(x = NULL)
}
\arguments{
\item{x}{a vector of counts.}
}
\description{
Calculate gini index.
}
