% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_changes.R
\name{calculate_lfc}
\alias{calculate_lfc}
\title{Calculate log fold changes.}
\usage{
calculate_lfc(
  data = NULL,
  id_column = 1,
  gene_column = 2,
  control_indices = NULL,
  treatment_indices = NULL,
  pseudocount = 0.5
)
}
\arguments{
\item{data}{sample count matrix.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{control_indices}{vector indices of columns containing control sample counts.}

\item{treatment_indices}{vector indices of columns containing treatment sample counts.}

\item{pseudocount}{a pseudocount (Default: 0.5).}
}
\value{
a data frame containing sample log fold changes.
}
\description{
Calculate log fold changes.
}
\details{
Takes the per-guide mean of control samples, adds a pseduocount to
all counts (including control mean) and calculates the log2 fold change.
Requires column indices be defined:
\itemize{
\item \code{id_column} - column containing guide (sgRNA) identifiers (Default = 1).
\item \code{gene_column} - column containing gene symbols/identifiers (Default = 2).
\item \code{control_indices} - indices of columns containing control sample counts.
\item \code{treatment_indices} - indices of columns containing treatment sample counts.
}
}
