% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{calculate_lfc_options}
\alias{calculate_lfc_options}
\title{Optparse options for output paths}
\usage{
calculate_lfc_options()
}
\value{
list of optparse options
}
\description{
Optparse options for output paths
}
