% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{check_dataframe}
\alias{check_dataframe}
\title{Check dataframe is not empty and indices are valid}
\usage{
check_dataframe(
  data = NULL,
  indices = NULL,
  check_na = FALSE,
  check_nan = FALSE
)
}
\arguments{
\item{data}{a dataframe.}

\item{indices}{vector of indices to check.}

\item{check_na}{whether to check for NA values.}

\item{check_nan}{whether to check for NaN values.}
}
\description{
Check dataframe is not empty and indices exist.
}
