% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{check_directory}
\alias{check_directory}
\title{Run basic checks on directory}
\usage{
check_directory(directory = NULL, ignore_empty = FALSE)
}
\arguments{
\item{directory}{the path of the directory which is being checked.}

\item{ignore_empty}{whether to skip check for empty directory}
}
\value{
Boolean
}
\description{
Checks that path is a directory, that the directory exists,
the directory can be read and that the directory is not empty.
}
