% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{check_file}
\alias{check_file}
\title{Run basic checks on input file}
\usage{
check_file(file = NULL, ignore_empty = FALSE)
}
\arguments{
\item{file}{the path of the file which is being checked.}

\item{ignore_empty}{whether to ignore empty file check.}
}
\value{
Boolean
}
\description{
Checks that path is a file, that the file exists,
the file can be read and that the file is not empty.
}
