% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{check_is_numeric_and_is_integer}
\alias{check_is_numeric_and_is_integer}
\title{Check value is numeric and is an integer.}
\usage{
check_is_numeric_and_is_integer(x)
}
\arguments{
\item{x}{a value.}
}
\value{
logical.
}
\description{
Check value is numeric and is an integer.
}
