% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{check_option}
\alias{check_option}
\title{Check optparse option is not null}
\usage{
check_option(option, value)
}
\arguments{
\item{option}{name of option to check}

\item{value}{value of option to check}
}
\value{
logical
}
\description{
Check optparse option is not null
}
