% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{compare_count_matrix_to_library}
\alias{compare_count_matrix_to_library}
\title{Compare count matrix to LibraryAnnotation.}
\usage{
compare_count_matrix_to_library(
  count_matrix = NULL,
  id_column = 1,
  gene_column = 2,
  library_annotation_object = NULL
)
}
\arguments{
\item{count_matrix}{data frame of sample counts.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{library_annotation_object}{LibraryAnnotations object}
}
\value{
logical.
}
\description{
Compares guide identifier and gene symbols in count matrix and \code{LibraryAnnotation} objects.
}
\seealso{
\link[rcrispr]{LibraryAnnotations-class}
}
