% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{compare_counts_to_library}
\alias{compare_counts_to_library}
\title{Compare SampleCounts to LibraryAnnotation.}
\usage{
compare_counts_to_library(
  sample_counts_object = NULL,
  library_annotation_object = NULL
)
}
\arguments{
\item{sample_counts_object}{SampleCounts object.}

\item{library_annotation_object}{LibraryAnnotations object.}
}
\value{
logical.
}
\description{
Compares guide identifier and gene symbols in \code{SampleCounts} and \code{LibraryAnnotation} objects.
}
\seealso{
\link[rcrispr]{SampleCounts-class}

\link[rcrispr]{LibraryAnnotations-class}
}
