% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{compare_matrix_to_sample_metadata}
\alias{compare_matrix_to_sample_metadata}
\title{Compare matrix to SampleMetadata.}
\usage{
compare_matrix_to_sample_metadata(
  data = NULL,
  id_column = 1,
  gene_column = NULL,
  sample_columns = 3,
  sample_metadata_object = NULL
)
}
\arguments{
\item{data}{data frame of sample counts or log fold changes.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{sample_columns}{indices of columns containing counts.}

\item{sample_metadata_object}{SampleMetadata object}
}
\value{
logical.
}
\description{
Checks sample names in matrix are present in \code{SampleMetadata} object.
}
\seealso{
\link[rcrispr]{SampleMetadata-class}
}
