% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{convert_variable_to_integer}
\alias{convert_variable_to_integer}
\title{Converts variable value to integers}
\usage{
convert_variable_to_integer(x = NULL)
}
\arguments{
\item{x}{input value}
}
\description{
Converts variable value to integer (where necessary)
}
