% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{convert_variable_to_numeric}
\alias{convert_variable_to_numeric}
\title{Converts variable value to numerics}
\usage{
convert_variable_to_numeric(x = NULL)
}
\arguments{
\item{x}{input value}
}
\description{
Converts variable value to numeric (where necessary)
}
