% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{correlation_options}
\alias{correlation_options}
\title{Optparse options for sample correlations}
\usage{
correlation_options()
}
\value{
list of optparse options
}
\description{
Optparse options for sample correlations
}
