% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{count_column_index_options}
\alias{count_column_index_options}
\title{Optparse options for count column indices}
\usage{
count_column_index_options()
}
\value{
list of optparse options
}
\description{
Optparse options for count column indices
}
