% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{count_library_outfile_options}
\alias{count_library_outfile_options}
\title{Optparse options for count and library output paths}
\usage{
count_library_outfile_options()
}
\value{
list of optparse options
}
\description{
Optparse options for count and library output paths
}
