% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_count_statistics.R
\name{count_matrix_stats}
\alias{count_matrix_stats}
\title{Build count matrix statistics}
\usage{
count_matrix_stats(
  count_matrix = NULL,
  id_column = 1,
  gene_column = 2,
  count_column = 3,
  low_counts = 30,
  total_reads = NULL
)
}
\arguments{
\item{count_matrix}{a count matrix data frame.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{count_column}{the index of column containing counts.}

\item{low_counts}{a threshold below which counts are considered low (default: 30).}

\item{total_reads}{a vector of read totals with samples as names.}
}
\description{
Calculate common CRISPR statistics from a count matrix
}
\details{
Optionally, total read counts can be given as a vector to \code{total_reads} with sample names as names for adding total reads (\code{total_reads}), proportion of mapped reads (\code{prop_mapped_reads}) and percentage of mapped reads (\code{pct_mapped_reads}).
}
