% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{count_path_options}
\alias{count_path_options}
\title{Optparse options for count paths}
\usage{
count_path_options()
}
\value{
list of optparse options
}
\description{
Optparse options for count paths
}
