% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{count_skip_options}
\alias{count_skip_options}
\title{Optparse options for skipping lines when reading files}
\usage{
count_skip_options()
}
\value{
list of optparse options
}
\description{
Optparse options for skipping lines when reading files
}
