% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{count_type_options}
\alias{count_type_options}
\title{Optparse options for input count type}
\usage{
count_type_options()
}
\value{
list of optparse options
}
\description{
Optparse options for input count type
}
