% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSampleCounts.R
\name{counts,SampleCounts-method}
\alias{counts,SampleCounts-method}
\title{Get counts from SampleCounts object}
\usage{
\S4method{counts}{SampleCounts}(object, processed = FALSE)
}
\arguments{
\item{object}{SampleCounts.}

\item{processed}{raw (TRUE) or processed counts (FALSE)}
}
\value{
dataframe containing sample counts
}
\description{
Get counts from SampleCounts object
}
