% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{crisprcleanr_normalisation_options}
\alias{crisprcleanr_normalisation_options}
\title{Optparse options for CRISPRcleanR normalisation}
\usage{
crisprcleanr_normalisation_options()
}
\value{
list of optparse options
}
\description{
Optparse options for CRISPRcleanR normalisation
}
