% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{crisprcleanr_output_options}
\alias{crisprcleanr_output_options}
\title{Optparse options for CRISPRcleanR outputs}
\usage{
crisprcleanr_output_options()
}
\value{
list of optparse options
}
\description{
Optparse options for CRISPRcleanR outputs
}
