% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{duplicate_guide_options}
\alias{duplicate_guide_options}
\title{Optparse options for output paths}
\usage{
duplicate_guide_options()
}
\value{
list of optparse options
}
\description{
Optparse options for output paths
}
