% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{filter_by_index_options}
\alias{filter_by_index_options}
\title{Optparse options for filtering data by index}
\usage{
filter_by_index_options()
}
\value{
list of optparse options
}
\description{
Optparse options for filtering data by index
}
