% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{filtered_guide_output_options}
\alias{filtered_guide_output_options}
\title{Optparse options for raw guide filter output}
\usage{
filtered_guide_output_options()
}
\value{
list of optparse options
}
\description{
Optparse options for raw guide filter output
}
