% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{fold_change_format_options}
\alias{fold_change_format_options}
\title{Optparse options for count formatting}
\usage{
fold_change_format_options()
}
\value{
list of optparse options
}
\description{
Optparse options for count formatting
}
