% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{fold_change_path_options}
\alias{fold_change_path_options}
\title{Optparse options for fold change paths}
\usage{
fold_change_path_options()
}
\value{
list of optparse options
}
\description{
Optparse options for fold change paths
}
