% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mageck_rra.R
\name{gene_summary_top_n_genes}
\alias{gene_summary_top_n_genes}
\title{Retrieve top n genes from MAGeCK gene summary}
\usage{
gene_summary_top_n_genes(
  gene_summary_object = NULL,
  n = 10,
  fdr = 0.05,
  direction = "neg"
)
}
\arguments{
\item{gene_summary_object}{a \code{MAGeCKGeneRRA} object.}

\item{n}{maximum number of genes to retrieve}

\item{fdr}{maximum FDR threshold}

\item{direction}{direction of results to retrieve (pos or neg)}
}
\value{
a data frame containing MAGeCK RRA gene results.
}
\description{
Collects top n genes from a \code{MageckGeneRRA} object.
}
