% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel_classification.R
\name{get_bagel_statistics}
\alias{get_bagel_statistics}
\title{Prepare BAGEL statistics}
\usage{
get_bagel_statistics(data = NULL, is_gene = FALSE, is_fc = FALSE)
}
\arguments{
\item{data}{data frame.}

\item{is_gene}{whether input data is at gene-level.}

\item{is_fc}{whether input data is at fc-level.}
}
\value{
logical.
}
\description{
Prepare BAGEL statistics
}
