% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{get_column_indices}
\alias{get_column_indices}
\title{Converts variable value to numerics}
\usage{
get_column_indices(n = NULL, df = NULL, sorted = FALSE)
}
\arguments{
\item{n}{column names}

\item{df}{data frame}

\item{sorted}{sort output list of columns}
}
\description{
Converts variable value to numeric (where necessary)
}
