% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{get_file_class}
\alias{get_file_class}
\title{Get file class}
\usage{
get_file_class(file = NULL)
}
\arguments{
\item{file}{the path of the file whose class is being determined.}
}
\value{
The class of \code{file}.
}
\description{
Returns the class of a given file.
}
