% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{get_guides_failing_filter}
\alias{get_guides_failing_filter}
\title{Identify guides failing count filter.}
\usage{
get_guides_failing_filter(
  count_matrix = NULL,
  id_column = 1,
  count_column = NULL,
  filter_indices = NULL,
  filter_method = "all",
  min_reads = 30
)
}
\arguments{
\item{count_matrix}{count matrix.}

\item{id_column}{index of column containing unique sgRNA identifiers.}

\item{count_column}{indices of columns containing counts.}

\item{filter_indices}{indices of columns on which to apply filter.}

\item{filter_method}{filter method from one of: all, any, mean or median (Default: all).}

\item{min_reads}{minimum number of reads for filter (Default: 30).}
}
\value{
dataframe
}
\description{
Identify guides which fail count filter.
}
