% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_annotations.R
\name{get_guides_with_no_coordinates}
\alias{get_guides_with_no_coordinates}
\title{Identify guides with no coordinates from library annotation object.}
\usage{
get_guides_with_no_coordinates(library_annotations_object = NULL)
}
\arguments{
\item{library_annotations_object}{LibraryAnnotations object.}
}
\value{
LibraryAnnotations object
}
\description{
Identify guides with no coordinates from a \code{LibraryAnnotations} object.
}
\seealso{
\link[rcrispr]{LibraryAnnotations-class}
}
