% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLibraryAnnotations.R
\name{get_library_annotations,LibraryAnnotations-method}
\alias{get_library_annotations,LibraryAnnotations-method}
\title{Get library annotations from LibraryAnnotations object}
\usage{
\S4method{get_library_annotations}{LibraryAnnotations}(
  object,
  processed = FALSE,
  sort_ids = FALSE,
  crisprcleanr = FALSE
)
}
\arguments{
\item{object}{LibraryAnnotations}

\item{processed}{raw (TRUE) or processed annotations (FALSE)}

\item{sort_ids}{order annotations by guide - only applicable when processed = TRUE (FALSE)}

\item{crisprcleanr}{format for CRISPRcleanR (TRUE) or default to processed annotations (FALSE)}
}
\value{
dataframe containing library annotations
}
\description{
Get library annotations from LibraryAnnotations object
}
