% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLibraryAnnotations.R
\name{get_library_annotations}
\alias{get_library_annotations}
\title{Generic for get_library_annotations method}
\usage{
get_library_annotations(
  object,
  processed = FALSE,
  sort_ids = FALSE,
  crisprcleanr = FALSE
)
}
\arguments{
\item{object}{object}

\item{processed}{logical}

\item{sort_ids}{logical}

\item{crisprcleanr}{logical}
}
\description{
Generic for get_library_annotations method
}
