% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMageckGeneRRA.R
\name{get_mageck_gene_summary,MageckGeneRRA-method}
\alias{get_mageck_gene_summary,MageckGeneRRA-method}
\title{Get gene summary from MageckGeneRRA object}
\usage{
\S4method{get_mageck_gene_summary}{MageckGeneRRA}(object, filters = NULL)
}
\arguments{
\item{object}{MageckGeneRRA}

\item{filters}{filter as string (e.g. 'neg.fdr < 0.05')}
}
\value{
data frame of MAGeCK gene summary results
}
\description{
Get gene summary from MageckGeneRRA object
}
