% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMageckSgrnaRRA.R
\name{get_mageck_sgrna_gene_results,MageckSgrnaRRA-method}
\alias{get_mageck_sgrna_gene_results,MageckSgrnaRRA-method}
\title{Get sgRNA gene results from MageckSgrnaRRA object}
\usage{
\S4method{get_mageck_sgrna_gene_results}{MageckSgrnaRRA}(object, gene = NULL)
}
\arguments{
\item{object}{MageckSgrnaRRA}

\item{gene}{name of gene}
}
\value{
data frame of MAGeCK sgRNA summary results
}
\description{
Get sgRNA gene results from MageckSgrnaRRA object
}
