% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMageckSgrnaRRA.R
\name{get_mageck_sgrna_gene_results}
\alias{get_mageck_sgrna_gene_results}
\title{Generic for get_mageck_sgrna_gene_results method}
\usage{
get_mageck_sgrna_gene_results(object, gene = NULL)
}
\arguments{
\item{object}{object}

\item{gene}{character}
}
\description{
Generic for get_mageck_sgrna_gene_results method
}
