% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMageckSgrnaRRA.R
\name{get_mageck_sgrna_summary,MageckSgrnaRRA-method}
\alias{get_mageck_sgrna_summary,MageckSgrnaRRA-method}
\title{Get sgRNA summary from MageckSgrnaRRA object}
\usage{
\S4method{get_mageck_sgrna_summary}{MageckSgrnaRRA}(object, filters = NULL)
}
\arguments{
\item{object}{MageckSgrnaRRA}

\item{filters}{filter as string (e.g. 'neg.fdr < 0.05')}
}
\value{
data frame of MAGeCK sgRNA summary results
}
\description{
Get sgRNA summary from MageckSgrnaRRA object
}
