% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMageckSgrnaRRA.R
\name{get_mageck_sgrna_summary}
\alias{get_mageck_sgrna_summary}
\title{Generic for get_mageck_sgrna_summary method}
\usage{
get_mageck_sgrna_summary(object, filters = NULL)
}
\arguments{
\item{object}{object}

\item{filters}{character}
}
\description{
Generic for get_mageck_sgrna_summary method
}
