% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSampleMetadata.R
\name{get_sample_metadata,SampleMetadata-method}
\alias{get_sample_metadata,SampleMetadata-method}
\title{Get sample metadata from SampleMetadata object}
\usage{
\S4method{get_sample_metadata}{SampleMetadata}(object, processed = FALSE)
}
\arguments{
\item{object}{SampleMetadata}

\item{processed}{raw (TRUE) or processed sample metadata (FALSE)}
}
\value{
dataframe containing library annotations
}
\description{
Get sample metadata from SampleMetadata object
}
