% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSampleMetadata.R
\name{get_sample_metadata}
\alias{get_sample_metadata}
\title{Generic for get_sample_metadata method}
\usage{
get_sample_metadata(object, processed = FALSE)
}
\arguments{
\item{object}{object}

\item{processed}{logical}
}
\description{
Generic for get_sample_metadata method
}
