% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{infile_column_index_options}
\alias{infile_column_index_options}
\title{Optparse options for infile column indices}
\usage{
infile_column_index_options()
}
\value{
list of optparse options
}
\description{
Optparse options for infile column indices
}
