% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{infile_format_options}
\alias{infile_format_options}
\title{Optparse options for infile formatting}
\usage{
infile_format_options()
}
\value{
list of optparse options
}
\description{
Optparse options for infile formatting
}
