% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{intermediate_qc_options}
\alias{intermediate_qc_options}
\title{Optparse options for intermediate qc}
\usage{
intermediate_qc_options()
}
\value{
list of optparse options
}
\description{
Optparse options for intermediate qc
}
