% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{library_annotation_column_index_options}
\alias{library_annotation_column_index_options}
\title{Optparse options for library column indices}
\usage{
library_annotation_column_index_options()
}
\value{
list of optparse options
}
\description{
Optparse options for library column indices
}
