% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{library_annotation_options}
\alias{library_annotation_options}
\title{Main optparse options for library annotations}
\usage{
library_annotation_options()
}
\value{
list of optparse options
}
\description{
Main optparse options for library annotations
}
