% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLibraryAnnotations.R
\name{library_has_coordinates,LibraryAnnotations-method}
\alias{library_has_coordinates,LibraryAnnotations-method}
\title{Check whether LibraryAnnotations object has coordinates}
\usage{
\S4method{library_has_coordinates}{LibraryAnnotations}(object)
}
\arguments{
\item{object}{LibraryAnnotations}
}
\value{
logical
}
\description{
Check whether LibraryAnnotations object has coordinates
}
