% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLibraryAnnotations.R
\name{library_has_coordinates}
\alias{library_has_coordinates}
\title{Generic for library_has_coordinates method}
\usage{
library_has_coordinates(object)
}
\arguments{
\item{object}{object}
}
\description{
Generic for library_has_coordinates method
}
