% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLibraryAnnotations.R
\docType{class}
\name{LibraryAnnotations-class}
\alias{LibraryAnnotations-class}
\title{An S4 class to represent library annotations}
\description{
An S4 class to represent library annotations
}
\section{Slots}{

\describe{
\item{\code{filepath}}{character.}

\item{\code{id_column}}{numeric.}

\item{\code{gene_column}}{numeric}

\item{\code{chr_column}}{character.}

\item{\code{chr_start_column}}{numeric}

\item{\code{chr_end_column}}{numeric}

\item{\code{file_separator}}{character.}

\item{\code{file_header}}{logical.}

\item{\code{annotations}}{data.frame.}
}}

