% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_count_statistics.R
\name{low_counts_per_sample}
\alias{low_counts_per_sample}
\title{Low counts per sample}
\usage{
low_counts_per_sample(
  count_matrix = NULL,
  id_column = 1,
  gene_column = 2,
  count_column = 3,
  filter_label = "zero_sgrnas",
  threshold = 1
)
}
\arguments{
\item{count_matrix}{a count matrix data frame.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{count_column}{the index of column containing counts.}

\item{filter_label}{name of low count column being returned.}

\item{threshold}{value below which counts are considered low (default: 1).}
}
\description{
Calculate number of low count guides per sample.
}
