% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{mageck_rra_summary_options}
\alias{mageck_rra_summary_options}
\title{Optparse options for MAGeCK RRA gene summary}
\usage{
mageck_rra_summary_options()
}
\value{
list of optparse options
}
\description{
Optparse options for MAGeCK RRA gene summary
}
