% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMageckGeneRRA.R
\docType{class}
\name{MageckGeneRRA-class}
\alias{MageckGeneRRA-class}
\title{An S4 class to represent MAGeCK RRA gene summary}
\description{
An S4 class to represent MAGeCK RRA gene summary
}
\section{Slots}{

\describe{
\item{\code{filepath}}{character.}

\item{\code{gene_summary}}{data.frame.}
}}

