% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMageckSgrnaRRA.R
\docType{class}
\name{MageckSgrnaRRA-class}
\alias{MageckSgrnaRRA-class}
\title{An S4 class to represent MAGeCK RRA sgRNA summary}
\description{
An S4 class to represent MAGeCK RRA sgRNA summary
}
\section{Slots}{

\describe{
\item{\code{filepath}}{character.}

\item{\code{sgrna_summary}}{data.frame.}
}}

