% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_count_plots.R
\name{plot_common_barplot}
\alias{plot_common_barplot}
\title{Generate bar plot}
\usage{
plot_common_barplot(df = NULL, ycol = NULL, ylab = NULL)
}
\arguments{
\item{df}{a count statisitic summary produced by \code{count_matrix_stats} with read mapping stats.}

\item{ycol}{column from count statistics to plot}

\item{ylab}{label for y axis}
}
\description{
Generate bar plot.
}
