% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_qc_plots.R
\name{plot_common_density_ridges}
\alias{plot_common_density_ridges}
\title{Generate ridgeline density plot}
\usage{
plot_common_density_ridges(df = NULL, xcol = NULL, xlab = NULL)
}
\arguments{
\item{df}{a data frame.}

\item{xcol}{column from data to plot}

\item{xlab}{label for x axis}
}
\description{
Generate ridgeline density plot.
}
