% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Generate correlation plot}
\usage{
plot_correlation(
  df = NULL,
  cor_columns = NULL,
  group_column = NULL,
  method = "pearson"
)
}
\arguments{
\item{df}{a data frame.}

\item{cor_columns}{list of columns to correlate.}

\item{group_column}{name of column containing groups (maximum of 3 groups).}

\item{method}{correlation method}
}
\description{
Generate correlation plot.
}
