% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mageck_rra.R
\name{plot_mageck_rra_gene_volcano}
\alias{plot_mageck_rra_gene_volcano}
\title{Plot MAGeCK RRA gene volcano}
\usage{
plot_mageck_rra_gene_volcano(
  gene_summary_object = NULL,
  n = 10,
  color_enriched = "royalblue4",
  color_depleted = "orangered4",
  fdr = 0.05
)
}
\arguments{
\item{gene_summary_object}{a \code{MAGeCKGeneRRA} object.}

\item{n}{maximum number of genes to label (in each direction)}

\item{color_enriched}{color for enriched genes}

\item{color_depleted}{color for depleted genes}

\item{fdr}{maximum FDR threshold}
}
\value{
a ggplot2 object.
}
\description{
Volcano plot of MAGeCK RRA gene summary results.
}
