% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mageck_rra.R
\name{plot_mageck_rra_sgrna_barplot}
\alias{plot_mageck_rra_sgrna_barplot}
\title{Plot MAGeCK RRA sgRNA LFC barplot}
\usage{
plot_mageck_rra_sgrna_barplot(sgrna_summary_object = NULL, gene = NULL)
}
\arguments{
\item{sgrna_summary_object}{a \code{MAGeCKGeneRRA} object.}

\item{gene}{name of gene to plot}
}
\value{
a ggplot2 object.
}
\description{
Barplot of MAGeCK RRA sgRNA log fold changes (LFC) for a given gene.
}
