% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_count_plots.R
\name{plot_mapping_statistics}
\alias{plot_mapping_statistics}
\title{Plot read mapping statistics}
\usage{
plot_mapping_statistics(df = NULL)
}
\arguments{
\item{df}{a count statisitic summary produced by \code{count_matrix_stats} with read mapping stats.}
}
\description{
Plot read mapping statistics from count statistic data frame.
}
