% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Generate PCA plot.}
\usage{
plot_pca(df = NULL, pc_x = "PC1", pc_y = "PC2")
}
\arguments{
\item{df}{a data frame.}

\item{pc_x}{principal component to plot on x axis}

\item{pc_y}{principal component to plot on y axis}
}
\description{
Generate PCA plot.
}
