% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel_classification.R
\name{prepare_essentiality_data}
\alias{prepare_essentiality_data}
\title{Prepare BAGEL statistics}
\usage{
prepare_essentiality_data(
  data = NULL,
  gene_column = 1,
  data_column = 2,
  classification_column = 3,
  threshold = 0.05
)
}
\arguments{
\item{data}{data frame.}

\item{gene_column}{index of column containing gene names.}

\item{data_column}{index of column containing values.}

\item{classification_column}{index of column containing BAGEL classifications.}
}
\value{
dataframe
}
\description{
Prepare BAGEL statistics
}
