% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{prepare_filepath}
\alias{prepare_filepath}
\title{Prepare file path}
\usage{
prepare_filepath(
  outfile = NULL,
  outdir = NULL,
  prefix = NULL,
  suffix = NULL,
  create_dir = FALSE
)
}
\arguments{
\item{outfile}{the name of the output file.}

\item{outdir}{the name of the output directory (defaults to current directory).}

\item{prefix}{a prefix to be added to the file name.}

\item{suffix}{a suffix to be added to the file name (before file extension).}

\item{create_dir}{whether to create directory if it does not exist}
}
\value{
A file path.
}
\description{
Prepares a file path by setting the output directory and modifying with a prefix and/or suffix.
}
