% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prepare_pca}
\alias{prepare_pca}
\title{Prepare data for PCA plot}
\usage{
prepare_pca(
  df = NULL,
  scaled = FALSE,
  transform = FALSE,
  log_transform = FALSE
)
}
\arguments{
\item{df}{a data frame.}

\item{scaled}{whether to scale data.}

\item{transform}{whether to transform data.}

\item{log_transform}{whether to log transform data.}
}
\description{
Prepare data for PCA plot.
}
