% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{process_column_indices}
\alias{process_column_indices}
\title{Process column indices into vector.}
\usage{
process_column_indices(columns = NULL)
}
\arguments{
\item{columns}{character string of column indices.}
}
\value{
a numeric vector of column indices.
}
\description{
Convert a character string of column indices into a numeric vector.
}
\examples{
process_column_indices('2,3,4,5')
process_column_indices('2:5')
process_column_indices('2-5')
process_column_indices('2:4,5')

}
