% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel_classification.R
\name{process_roc}
\alias{process_roc}
\title{Process ROC}
\usage{
process_roc(roc_obj = NULL, predictor_min = NULL, threshold = 0.05)
}
\arguments{
\item{roc_obj}{pROC ROC object.}

\item{predictor_min}{minimum value of predictor.}

\item{threshold}{threshold value.}
}
\value{
dataframe
}
\description{
Process ROC
}
