% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{read_count_matrix_file}
\alias{read_count_matrix_file}
\title{Read in sample count matrix.}
\usage{
read_count_matrix_file(
  filepath = NULL,
  file_separator = "\\t",
  file_header = TRUE,
  id_column = 1,
  gene_column = 2,
  count_column = NULL,
  count_index_increment = 0,
  processed = FALSE,
  ...
)
}
\arguments{
\item{filepath}{Character string specifying a file path.}

\item{file_separator}{count file separator.}

\item{file_header}{count file header}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{count_column}{vector indices of columns containing sample counts.}

\item{processed}{logical of whether to apply predefined column names.}

\item{...}{additional read.delim parameters.}
}
\value{
a data frame containing sample counts.
}
\description{
Read file containing count matrix into a data frame.
}
\details{
Reads a count matrix file, as defined by \code{filepath}, into a data frame.
Requires two annotation column indices be defined:
\itemize{
\item \code{id_column} - column containing guide (sgRNA) identifiers (Default = 1).
\item \code{gene_column} - column containing gene symbols/identifiers (Default = 2).
\item \code{count_column} - indices of columns containing sample counts.
}
There must also be one or more columns containing sample counts whose indices are
indicated using \code{count_column}. If \code{count_column} is \code{NULL} then it is assumed
that all columns except \code{id_column} and \code{gene_column} contain count data.

Assumes by default that the count file has a header, as defined by \code{file_header},
and that it is tab-delimited, as defined by \code{file_separator}.
}
