% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{read_file_to_dataframe}
\alias{read_file_to_dataframe}
\title{Read delimited file}
\usage{
read_file_to_dataframe(
  filepath = NULL,
  file_separator = "\\t",
  file_header = T,
  column_indices = NULL,
  ...
)
}
\arguments{
\item{filepath}{delimited file to be read.}

\item{file_separator}{field separator.}

\item{file_header}{logical stating whether file has header.}

\item{column_indices}{columns to include}

\item{...}{additional read.delim parameters.}
}
\value{
dataframe
}
\description{
Read delimited file
}
