% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_annotations.R
\name{read_library_annotation_file}
\alias{read_library_annotation_file}
\title{Read library annotation file.}
\usage{
read_library_annotation_file(
  filepath = NULL,
  id_column = 1,
  gene_column = 2,
  chr_column = NULL,
  chr_start_column = NULL,
  chr_end_column = NULL,
  file_separator = "\\t",
  file_header = TRUE,
  ...
)
}
\arguments{
\item{filepath}{character string specifying a file path.}

\item{id_column}{index of column containing unique sgRNA identifiers.}

\item{gene_column}{index of column containing gene symbols.}

\item{chr_column}{index of column containing chromosome name.}

\item{chr_start_column}{index of column containing chromosome start position.}

\item{chr_end_column}{index of column containing chromosome position.}

\item{file_separator}{library annotation file separator.}

\item{file_header}{whether library annotation file contains a header.}

\item{...}{additional read.delim parameters.}
}
\value{
a data frame containing library annotations.
}
\description{
Reads in a library annotation file to a \code{LibraryAnnotations} object.
}
