% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mageck_rra.R
\name{read_mageck_rra_gene_summary}
\alias{read_mageck_rra_gene_summary}
\title{Reads MAGeCK RRA gene summary}
\usage{
read_mageck_rra_gene_summary(filepath = NULL, ...)
}
\arguments{
\item{filepath}{character string specifying a file path.}

\item{...}{additional read.delim parameters.}
}
\value{
a data frame containing MAGeCK RRA gene summary.
}
\description{
Reads in a MAGeCK RRA gene summary file to a \code{MageckGeneRRA} object.
}
