% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mageck_rra.R
\name{read_mageck_rra_sgrna_summary}
\alias{read_mageck_rra_sgrna_summary}
\title{Reads MAGeCK RRA sgRNA summary}
\usage{
read_mageck_rra_sgrna_summary(filepath = NULL, ...)
}
\arguments{
\item{filepath}{character string specifying a file path.}

\item{...}{additional read.delim parameters.}
}
\value{
a data frame containing MAGeCK RRA sgRNA summary.
}
\description{
Reads in a MAGeCK RRA sgRNA summary file to a \code{MageckSgrnaRRA} object.
}
