% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{read_sample_count_file}
\alias{read_sample_count_file}
\title{Read sample count file.}
\usage{
read_sample_count_file(
  filepath = NULL,
  sample_name = NULL,
  id_column = 1,
  gene_column = 2,
  count_column = 3,
  file_separator = "\\t",
  file_header = TRUE,
  strip_ids = FALSE,
  ...
)
}
\arguments{
\item{filepath}{character string specifying a file path.}

\item{sample_name}{the sample name.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{count_column}{the index of column containing counts.}

\item{file_separator}{count file separator.}

\item{file_header}{whether count file(s) contain a header.}

\item{strip_ids}{whether to make syntactically valid id names.}

\item{...}{additional read.delim parameters.}
}
\value{
a \code{SampleCounts} object.
}
\description{
Reads in an individual sample count file to a \code{SampleCounts} object.
}
\details{
Reads a file containing guide counts for a single file and creates a
\link[rcrispr]{SampleCounts-class} object.

Expects at least three columns which are defined by their index:
\itemize{
\item \code{id_column} - column containing guide (sgRNA) identifiers (Default = 1).
\item \code{gene_column} - column containing gene symbols/identifiers (Default = 2).
\item \code{count_column} - column containing sample counts (Default = 3).
}
Assumes by default that the count file has a header, defined by \code{file_header},
and that it is tab-delimited, defined by \code{file_separator}.
}
\seealso{
\link[rcrispr]{SampleCounts-class}
}
