% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{read_sample_count_files}
\alias{read_sample_count_files}
\title{Read sample count files.}
\usage{
read_sample_count_files(
  count_directory = NULL,
  id_column = 1,
  gene_column = 2,
  count_column = 3,
  file_separator = "\\t",
  file_header = TRUE,
  sample_metadata_object = NULL,
  ...
)
}
\arguments{
\item{count_directory}{path of directory containing counts.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{count_column}{the index of column containing counts.}

\item{file_separator}{count file separator.}

\item{file_header}{whether count file(s) contain a header.}

\item{sample_metadata_object}{a sample metadata object}

\item{...}{additional read.delim parameters.}
}
\value{
a list of \code{SampleCounts} objects.
}
\description{
Reads in individual sample count files to a list of \code{SampleCounts} objects.
}
\details{
Using a list of file names from a \link[rcrispr]{SampleMetadata-class} object,
reads in individual sample counts to \link[rcrispr]{SampleCounts-class} objects.

Requires three count columns to be defined by their index:
\itemize{
\item \code{id_column} - column containing guide (sgRNA) identifiers (Default = 1).
\item \code{gene_column} - column containing gene symbols/identifiers (Default = 2).
\item \code{count_column} - column containing sample counts (Default = 3).
}
Assumes by default that the count file has a header, defined by \code{file_header},
and that it is tab-delimited, defined by \code{file_separator}.
}
\seealso{
\link[rcrispr]{SampleCounts-class}
}
