% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{read_sample_metadata_file}
\alias{read_sample_metadata_file}
\title{Read a sample metadata file.}
\usage{
read_sample_metadata_file(
  filepath = NULL,
  filename_column = 1,
  label_column = 2,
  plasmid_column = 3,
  control_column = 4,
  treatment_column = 5,
  group_column = NULL,
  reads_column = NULL,
  file_separator = "\\t",
  file_header = TRUE,
  ...
)
}
\arguments{
\item{filepath}{Character string specifying a file path.}

\item{filename_column}{the index of column containing sample count fileanames.}

\item{label_column}{the index of column containing sample names/labels.}

\item{plasmid_column}{the index of column showing whether sample is a plasmid sample.}

\item{control_column}{the index of column showing whether sample is a control sample.}

\item{treatment_column}{the index of column showing whether sample is a treatment sample.}

\item{group_column}{the index of column containing sample group.}

\item{reads_column}{the index of column containing total sequencing reads per sample.}

\item{file_separator}{count file separator.}

\item{file_header}{count file header.}

\item{...}{additional read.delim parameters.}
}
\value{
a \code{SampleMetadata} object..
}
\description{
Read in a sample metadata file to a \code{SampleMetadata} object.
}
\seealso{
\link[rcrispr]{SampleMetadata-class}
}
