% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{remove_guides_from_count_matrix}
\alias{remove_guides_from_count_matrix}
\title{Remove guides from count matrix.}
\usage{
remove_guides_from_count_matrix(
  count_matrix = NULL,
  id_column = 1,
  guides_to_remove = NULL
)
}
\arguments{
\item{count_matrix}{sample count matrix.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{guides_to_remove}{guide identifiers to remove.}
}
\value{
dataframe
}
\description{
Remove guides from a count matrix.
}
