% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_annotations.R
\name{remove_guides_from_library_annotations_object}
\alias{remove_guides_from_library_annotations_object}
\title{Remove guides from library annotation object.}
\usage{
remove_guides_from_library_annotations_object(
  library_annotations_object = NULL,
  guides_to_remove = NULL
)
}
\arguments{
\item{library_annotations_object}{LibraryAnnotations object.}

\item{guides_to_remove}{guide identifiers to remove.}
}
\value{
LibraryAnnotations object
}
\description{
Remove guides from a \code{LibraryAnnotations} object.
}
\seealso{
\link[rcrispr]{LibraryAnnotations-class}
}
