% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{remove_guides_from_sample_counts}
\alias{remove_guides_from_sample_counts}
\title{Remove guides from sample counts.}
\usage{
remove_guides_from_sample_counts(
  sample_counts_object = NULL,
  guides_to_remove = NULL
)
}
\arguments{
\item{sample_counts_object}{SampleCounts object.}

\item{guides_to_remove}{guide identifiers to remove.}
}
\value{
SampleCounts object
}
\description{
Remove guides from a \code{SampleCounts} object.
}
\seealso{
\link[rcrispr]{SampleCounts-class}
}
