% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{remove_no_coordinate_guide_options}
\alias{remove_no_coordinate_guide_options}
\title{Optparse options for removing guides with no coordinates}
\usage{
remove_no_coordinate_guide_options()
}
\value{
list of optparse options
}
\description{
Optparse options for removing guides with no coordinates
}
