% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\name{reorder_count_matrix_by_sample_type}
\alias{reorder_count_matrix_by_sample_type}
\title{Reorder count matrix by sample metadata.}
\usage{
reorder_count_matrix_by_sample_type(
  count_matrix = NULL,
  sample_metadata_object = NULL
)
}
\arguments{
\item{count_matrix}{count matrix data frame}

\item{sample_metadata_object}{SampleMetadata object}
}
\value{
count matrix data frame.
}
\description{
Reorders sample columns in count matrix using sample types (plasmid, control or treatment) from a \code{SampleMetadata} object.
}
\seealso{
\link[rcrispr]{SampleMetadata-class}
}
