% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{sample_metadata_sample_filename_column_index_options}
\alias{sample_metadata_sample_filename_column_index_options}
\title{Optparse options for sample metadata count files column indices}
\usage{
sample_metadata_sample_filename_column_index_options()
}
\value{
list of optparse options
}
\description{
Optparse options for sample metadata count files column indices
}
