% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSampleCounts.R
\docType{class}
\name{SampleCounts-class}
\alias{SampleCounts-class}
\title{An S4 class to represent a single set of sample counts.}
\description{
An S4 class to represent a single set of sample counts.
}
\section{Slots}{

\describe{
\item{\code{sample_name}}{character.}

\item{\code{filepath}}{character.}

\item{\code{id_column}}{numeric.}

\item{\code{gene_column}}{numeric.}

\item{\code{count_column}}{numeric.}

\item{\code{file_separator}}{character.}

\item{\code{file_header}}{logical.}

\item{\code{counts}}{data.frame.}
}}

