% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSampleMetadata.R
\docType{class}
\name{SampleMetadata-class}
\alias{SampleMetadata-class}
\title{An S4 class to represent sample metadata (mapping)}
\description{
An S4 class to represent sample metadata (mapping)
}
\section{Slots}{

\describe{
\item{\code{filepath}}{character.}

\item{\code{filename_column}}{numeric.}

\item{\code{label_column}}{numeric}

\item{\code{plasmid_column}}{numeric.}

\item{\code{control_column}}{numeric.}

\item{\code{treatment_column}}{numeric.}

\item{\code{group_column}}{numeric.}

\item{\code{reads_column}}{numeric.}

\item{\code{file_separator}}{character.}

\item{\code{file_header}}{logical.}

\item{\code{metadata}}{data.frame.}
}}

