% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{save_plot_list}
\alias{save_plot_list}
\title{Save list of plots}
\usage{
save_plot_list(
  plot_list = NULL,
  outdir = NULL,
  prefix = NULL,
  suffix = NULL,
  ...
)
}
\arguments{
\item{plot_list}{list of plots to save.}

\item{outdir}{the name of the output directory (defaults to current directory).}

\item{prefix}{a prefix to be added to the file name.}

\item{suffix}{a suffix to be added to the file name (before file extension).}

\item{...}{parameters for ggsave.}
}
\description{
Converts variable value to numeric (where necessary)
}
