% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{scaling_options}
\alias{scaling_options}
\title{Optparse options for scaling LFCs and BFs}
\usage{
scaling_options()
}
\value{
list of optparse options
}
\description{
Optparse options for scaling LFCs and BFs
}
