% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{sequencing_qc_options}
\alias{sequencing_qc_options}
\title{Optparse options for raw count QC}
\usage{
sequencing_qc_options()
}
\value{
list of optparse options
}
\description{
Optparse options for raw count QC
}
