% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helper.R
\name{strip_id_options}
\alias{strip_id_options}
\title{Optparse options for stripping guide ids}
\usage{
strip_id_options()
}
\value{
list of optparse options
}
\description{
Optparse options for stripping guide ids
}
