% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_count_statistics.R
\name{total_counts_per_sample}
\alias{total_counts_per_sample}
\title{Total counts per sample}
\usage{
total_counts_per_sample(
  count_matrix = NULL,
  id_column = 1,
  gene_column = 2,
  count_column = 3
)
}
\arguments{
\item{count_matrix}{a count matrix data frame.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}

\item{count_column}{the index of column containing counts.}
}
\description{
Calculate total counts per sample.
}
