% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{write_rdata_to_file}
\alias{write_rdata_to_file}
\title{Write R data to file.}
\usage{
write_rdata_to_file(
  data = NULL,
  outfile = NULL,
  outdir = NULL,
  prefix = NULL,
  suffix = NULL
)
}
\arguments{
\item{data}{list of objects to save.}

\item{outfile}{the name of the output file.}

\item{outdir}{the name of the output directory (defaults to current directory).}

\item{prefix}{a prefix to be added to the file name.}

\item{suffix}{a suffix to be added to the file name (before file extension).}
}
\value{
The class of \code{file}.
}
\description{
Writes R data objects to file.
}
