% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_changes.R
\name{calculate_gene_lfc}
\alias{calculate_gene_lfc}
\title{Calculate gene log fold changes.}
\usage{
calculate_gene_lfc(data = NULL, id_column = 1, gene_column = 2)
}
\arguments{
\item{data}{sample count matrix.}

\item{id_column}{the index of column containing unique sgRNA identifiers.}

\item{gene_column}{the index of column containing gene symbols.}
}
\value{
a data frame containing sample log fold changes.
}
\description{
Calculate gene log fold changes.
}
\details{
Takes the per-guide mean of control samples, adds a pseduocount to
all counts (including control mean) and calculates the log2 fold change.
Requires column indices be defined:
\itemize{
\item \code{id_column} - column containing guide (sgRNA) identifiers (Default = 1).
\item \code{gene_column} - column containing gene symbols/identifiers (Default = 2).
}
}
