% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSampleCounts.R
\name{counts}
\alias{counts}
\title{Generic for counts method}
\usage{
counts(object, processed = FALSE, sort_ids = FALSE)
}
\arguments{
\item{object}{object}

\item{processed}{logical}

\item{sort_ids}{logical}
}
\description{
Generic for counts method
}
