% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_count_plots.R
\name{plot_common_barplot}
\alias{plot_common_barplot}
\title{Generate bar plot}
\usage{
plot_common_barplot(
  df = NULL,
  xcol = NULL,
  xlab = NULL,
  ycol = NULL,
  ylab = NULL
)
}
\arguments{
\item{df}{a count statisitic summary produced by \code{count_matrix_stats} with read mapping stats.}

\item{xcol}{column from count statistics to plot on x axis}

\item{xlab}{label for x axis}

\item{ycol}{column from count statistics to plot on y axis}

\item{ylab}{label for y axis}
}
\description{
Generate bar plot.
}
