% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_qc_plots.R
\name{plot_common_violin}
\alias{plot_common_violin}
\title{Generate violin plot}
\usage{
plot_common_violin(df = NULL, ycol = NULL, ylab = NULL)
}
\arguments{
\item{df}{a count statisitic summary produced by \code{count_matrix_stats} with read mapping stats.}

\item{ycol}{column from data to plot}

\item{ylab}{label for y axis}
}
\description{
Generate violin plot.
}
